requirements.txt
nose
sphinx
kafka-python
avro-python3==1.9.0